fix(notebook): bnb pin, auto-phase, dtype kwarg — from first Kaggle run
First real-session feedback (Kaggle L4):

- bitsandbytes ImportError: transformers now requires >=0.46.1; our cell
  pinned >=0.43 AND suppressed the install error with 2>/dev/null. Pin
  bumped, error suppression removed, import + version verified in-cell.
- torch_dtype deprecation: switch to dtype= (notebook C1 cell + sft.py);
  transformers pin bumped to >=4.56 to match.
- PHASE now auto-selects (user UX request): C1 until the donor baseline
  report exists on the Hub, then C5. Manual editing of the config cell
  no longer required; AVA_PHASE env var overrides. SHARD_MINUTES removed
  from the notebook — hw_profile owns cadence.

+++ experiments/exp6_v3/train/sft.py
@@ -91,7 +91,7 @@ def _build_qlora_model(cfg: SFTConfig, compute_dtype: torch.dtype) -> tuple[Any,
         cfg.donor,
         quantization_config=bnb,
         device_map="auto",
-        torch_dtype=compute_dtype,
+        dtype=compute_dtype,  # torch_dtype deprecated in transformers >= 4.56
         trust_remote_code=True,
     )
     model = prepare_model_for_kbit_training(model, use_gradient_checkpointing=True)
